# contracts/openapi/mobile-v1.yaml
openapi: 3.1.0
info:
  title: 404 Mobile API
  version: 0.1.0
  description: >
    Modular monolith baseline contract for the 404 mobile application.
servers:
  - url: /api/v1
...
paths:
...
  /facades/notifications:
    get:
      tags: [Facade - Notification]
      summary: Get notification feed payload
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Notification feed payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationFacadeResponse'
...
  /auth/kakao/login:
    post:
      tags: [Auth]
      summary: Login with Kakao authorization code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KakaoLoginRequest'
      responses:
        '200':
          description: Login success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
  /auth/terms/accept:
    post:
      tags: [Auth]
      summary: Accept required terms
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AcceptTermsRequest'
      responses:
        '204':
          description: Accepted
  /auth/session/refresh:
    post:
      tags: [Auth]
      summary: Refresh session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshSessionRequest'
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshSessionResponse'
  /auth/logout:
    post:
      tags: [Auth]
      summary: Logout current session
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Logged out
...
  /houses/{houseId}/chores:
    parameters:
      - $ref: '#/components/parameters/HouseId'
    post:
      tags: [Chore]
      summary: Create chore rule
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateChoreRequest'
      responses:
        '201':
          description: Chore created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChoreRule'
  /houses/{houseId}/chores/today:
    parameters:
      - $ref: '#/components/parameters/HouseId'
    get:
      tags: [Chore]
      summary: List today chores
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: date
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Today chore list
          content:
            application/json:
              schema:
                type: object
                properties:
                  chores:
                    type: array
                    items:
                      $ref: '#/components/schemas/TodayChore'
  /chore-instances/{choreInstanceId}/completion:
    parameters:
      - $ref: '#/components/parameters/ChoreInstanceId'
    post:
      tags: [Chore]
      summary: Complete chore instance
      security:
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompleteChoreRequest'
      responses:
        '200':
          description: Chore completed
    delete:
      tags: [Chore]
      summary: Cancel chore completion
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Completion cancelled
  /houses/{houseId}/adjustments/substitute-requests:
    parameters:
      - $ref: '#/components/parameters/HouseId'
    post:
      tags: [Adjustment]
      summary: Request substitute
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubstituteRequest'
      responses:
        '201':
          description: Adjustment request created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdjustmentRequest'
  /adjustments/{adjustmentRequestId}/accept:
    parameters:
      - $ref: '#/components/parameters/AdjustmentRequestId'
    post:
      tags: [Adjustment]
      summary: Accept adjustment request
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Adjustment accepted
  /adjustments/{adjustmentRequestId}/reject:
    parameters:
      - $ref: '#/components/parameters/AdjustmentRequestId'
    post:
      tags: [Adjustment]
      summary: Reject adjustment request
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Adjustment rejected
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    HouseId:
      in: path
      name: houseId
      required: true
      schema:
        type: string
        format: uuid
    ChoreInstanceId:
      in: path
      name: choreInstanceId
      required: true
      schema:
        type: string
        format: uuid
    AdjustmentRequestId:
      in: path
      name: adjustmentRequestId
      required: true
      schema:
        type: string
        format: uuid
...
    Cursor:
      in: query
      name: cursor
      schema:
        type: string
    Limit:
      in: query
      name: limit
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  schemas:
...
    NotificationFacadeResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/NotificationItem'
        unreadCount:
          type: integer
        nextCursor:
          type: string
          nullable: true
...
    KakaoLoginRequest:
      type: object
      required: [authorizationCode, redirectUri]
      properties:
        authorizationCode:
          type: string
        redirectUri:
          type: string
        deviceId:
          type: string
        deviceName:
          type: string
    LoginResponse:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        isNewUser:
          type: boolean
        accessToken:
          type: string
        refreshToken:
          type: string
        requiredTerms:
          type: array
          items:
            type: string
    AcceptTermsRequest:
      type: object
      required: [terms]
      properties:
        terms:
          type: array
          items:
            type: object
            required: [type, version]
            properties:
              type:
                type: string
              version:
                type: string
    RefreshSessionRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    RefreshSessionResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
...
    CreateChoreRequest:
      type: object
      required: [spaceId, title]
      properties:
        spaceId:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
          nullable: true
        estimatedMinutes:
          type: integer
          nullable: true
        defaultAssigneeMembershipId:
          type: string
          format: uuid
          nullable: true
        recurrence:
          $ref: '#/components/schemas/Recurrence'
    ChoreRule:
      type: object
      properties:
        choreRuleId:
          type: string
          format: uuid
        spaceId:
          type: string
          format: uuid
        title:
          type: string
        status:
          type: string
        recurrence:
          $ref: '#/components/schemas/Recurrence'
    Recurrence:
      type: object
      properties:
        frequency:
          type: string
        interval:
          type: integer
        daysOfWeek:
          type: array
          items:
            type: string
        startDate:
          type: string
          format: date
          nullable: true
    TodayChore:
      type: object
      properties:
        choreInstanceId:
          type: string
          format: uuid
        title:
          type: string
        spaceName:
          type: string
        assigneeName:
          type: string
          nullable: true
        status:
          type: string
        scheduledDate:
          type: string
          format: date
...
    CompleteChoreRequest:
      type: object
      properties:
        memo:
          type: string
          nullable: true
        proofImageUrl:
          type: string
          nullable: true
    SubstituteRequest:
      type: object
      required: [choreInstanceId]
      properties:
        choreInstanceId:
          type: string
          format: uuid
        reason:
          type: string
          nullable: true
        expiresAt:
          type: string
          format: date-time
          nullable: true
    AdjustmentRequest:
      type: object
      properties:
        adjustmentRequestId:
          type: string
          format: uuid
        choreInstanceId:
          type: string
          format: uuid
        requestType:
          type: string
        status:
          type: string
        expiresAt:
          type: string
          format: date-time
          nullable: true
    NotificationItem:
      type: object
      properties:
        notificationId:
          type: string
          format: uuid
        type:
          type: string
        title:
          type: string
        body:
          type: string
        isRead:
          type: boolean
        occurredAt:
          type: string
          format: date-time
        payload:
          type: object
          additionalProperties: true
...